Move Cargo manifest path resolution to compilation.rs
